requirements.txt
numpy>=1
gitpython
# can comment out stuff here due to setup.py parser
#
#compliance-checker
